# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Orchestrated Study Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface StudySessionInput {
  /** @minLength 1 */
  title: string;
}

export interface StudySession {
  id: string;
  userId: string;
  title: string;
  createdAt: string;
  lastAccessed: string;
  documentCount: number;
  messageCount: number;
}

export interface Document {
  id: string;
  sessionId: string;
  filename: string;
  mimeType: string;
  content: string;
  uploadedAt: string;
}

export interface StudySessionWithDocuments {
  id: string;
  userId: string;
  title: string;
  createdAt: string;
  lastAccessed: string;
  /** @nullable */
  notes: string | null;
  documents: Document[];
}

export interface SessionNotesUpdate {
  notes: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: string;
  sessionId: string;
  /** @nullable */
  documentId: string | null;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ChatInput {
  sessionId: string;
  documentId: string;
  /** @minLength 1 */
  message: string;
  includeNotes?: boolean;
}

export interface DocumentUpload {
  file: Blob;
}

export interface DashboardSummary {
  totalSessions: number;
  totalDocuments: number;
  totalMessages: number;
  recentSessions: StudySession[];
}