# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Orchestrated Study Companion API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: sessions
    description: Study session operations
  - name: documents
    description: Document management operations
  - name: messages
    description: Chat message operations
  - name: chat
    description: AI chat operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List all study sessions for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new study session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudySessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a single study session with its documents
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySessionWithDocuments"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a study session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{id}/notes:
    patch:
      operationId: updateSessionNotes
      tags: [sessions]
      summary: Save personal notes for a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionNotesUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySessionWithDocuments"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document to a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "201":
          description: Uploaded document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a chat message and stream AI response (SSE)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream: {}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AuthInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt

    StudySessionInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
      required:
        - title

    StudySession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        lastAccessed:
          type: string
          format: date-time
        documentCount:
          type: integer
        messageCount:
          type: integer
      required:
        - id
        - userId
        - title
        - createdAt
        - lastAccessed
        - documentCount
        - messageCount

    StudySessionWithDocuments:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        lastAccessed:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required:
        - id
        - userId
        - title
        - createdAt
        - lastAccessed
        - notes
        - documents

    SessionNotesUpdate:
      type: object
      properties:
        notes:
          type: string
      required:
        - notes

    Document:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        content:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - filename
        - mimeType
        - content
        - uploadedAt

    Message:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        documentId:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - documentId
        - role
        - content
        - createdAt

    ChatInput:
      type: object
      properties:
        sessionId:
          type: string
        documentId:
          type: string
        message:
          type: string
          minLength: 1
        includeNotes:
          type: boolean
      required:
        - sessionId
        - documentId
        - message

    DocumentUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file

    DashboardSummary:
      type: object
      properties:
        totalSessions:
          type: integer
        totalDocuments:
          type: integer
        totalMessages:
          type: integer
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/StudySession"
      required:
        - totalSessions
        - totalDocuments
        - totalMessages
        - recentSessions