# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Orchestrated Study Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { StudySession } from './studySession';

export interface DashboardSummary {
  totalSessions: number;
  totalDocuments: number;
  totalMessages: number;
  recentSessions: StudySession[];
}