# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Orchestrated Study Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: string;
  userId: string;
  title: string;
  createdAt: Date;
  lastAccessed: Date;
  documentCount: number;
  messageCount: number;
}