# lib/api-zod/src/generated/types/studySessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Orchestrated Study Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface StudySessionInput {
  /** @minLength 1 */
  title: string;
}