# lib/api-zod/src/generated/types/studySessionWithDocuments.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Orchestrated Study Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface StudySessionWithDocuments {
  id: string;
  userId: string;
  title: string;
  createdAt: Date;
  lastAccessed: Date;
  /** @nullable */
  notes: string | null;
  documents: Document[];
}